+++ packages/next/shared/lib/head.tsx
@@ -5,7 +5,7 @@ import Effect from './side-effect'
 import { AmpStateContext } from './amp-context'
 import { HeadManagerContext } from './head-manager-context'
 import { isInAmpMode } from './amp-mode'
-import { warnOnce } from './utils'
+import { warnOnce } from './utils/warn-once'
 
 type WithInAmpMode = {
   inAmpMode?: boolean
