+++ packages/next/shared/lib/utils.ts
@@ -400,19 +400,6 @@ export async function loadGetInitialProps<
   return props
 }
 
-let warnOnce = (_: string) => {}
-if (process.env.NODE_ENV !== 'production') {
-  const warnings = new Set<string>()
-  warnOnce = (msg: string) => {
-    if (!warnings.has(msg)) {
-      console.warn(msg)
-    }
-    warnings.add(msg)
-  }
-}
-
-export { warnOnce }
-
 export const SP = typeof performance !== 'undefined'
 export const ST =
   SP &&
